requirements.txt
Flask>=2.0
pdfplumber>=0.7
pandas>=1.0
tabula-py>=2.0
requests>=2.28
